Revert "Moving the CurrentGameEndpoint and FeaturedGameEndpoint to a common SpectatorEndpoint."
This reverts commit 58fd437e8d3e36bb9e3ef3db95d20bda9801836a.
